+++ packages/orq-rc/src/models/operations/listdatasetdatapoints.ts
@@ -2604,7 +2604,7 @@ export const ListDatasetDatapointsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-08-18T10:59:53.795Z",
+    "2025-08-19T12:28:47.558Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -2668,7 +2668,7 @@ export const ListDatasetDatapointsData$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-08-18T10:59:53.795Z"))
+  updated: z.date().default(() => new Date("2025-08-19T12:28:47.558Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {