+++ packages/orq-rc/RELEASES.md
@@ -3119,3 +3119,13 @@ Based on:
 - [typescript v3.11.0-rc.22] packages/orq-rc
 ### Releases
 - [NPM v3.11.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/3.11.0-rc.22 - packages/orq-rc
+
+## 2025-08-19 12:30:16
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.606.0 (2.686.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.12.0-rc.0] packages/orq-rc
+### Releases
+- [NPM v3.12.0-rc.0] https://www.npmjs.com/package/@orq-ai/node/v/3.12.0-rc.0 - packages/orq-rc